api_requirements.txt
fastapi==0.115.12
joblib==1.5.1
kafka-python==2.2.11
lightgbm==4.6.0
numpy==2.2.6
pandas==2.3.0
prometheus_client==0.22.1
pydantic==2.11.5
pydantic_core==2.33.2
requests==2.32.3
scikit-learn==1.7.0
uvicorn==0.34.3
xgboost==3.0.2
